♻️ Refactor Home.jsx move to calendar component to  mainCalendarWrapper.jsx
